# model/src/main/java/com/showtime/model/view/sysauthority/SysAuthorityView.java
package com.showtime.model.view.sysauthority;

import com.showtime.model.entity.sysauthority.SysAuthority;
import io.swagger.annotations.ApiModel;

import java.io.Serializable;

/**
 * <b><code>SysAuthorityView</code></b>
 * <p/>
 * SysAuthorityView
 * <p/>
 * <b>Creation Time:</b> 2017/05/17 21:18.
 *
 * @author lvxin
 * @version 1.0.0
 * @since model 1.0.0
 */
@ApiModel
public class SysAuthorityView extends SysAuthority implements Serializable {

    /**
     * The constant serialVersionUID.
     */
    private static final long serialVersionUID = 8672084909411877404L;

}