# rs/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "RESTful API文档",
    "version" : "v1.0",
    "title" : "RESTful API文档",
    "termsOfService" : "http://www.showtime.com/",
    "contact" : {
      "name" : "showtime",
      "url" : "http://www.showtime.com",
      "email" : "594470882@qq.com"
    }
  },
  "basePath" : "/rs",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/api/v1/sysauthorities" : {
      "get" : {
        "tags" : [ "后台管理权限接口" ],
        "summary" : "获取后台管理权限列表",
        "description" : "通过查询条件获取后台管理权限列表",
        "operationId" : "getSysAuthoritys",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "description" : "权限",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "enabled",
          "in" : "query",
          "description" : "状态",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageNumber",
          "in" : "query",
          "description" : "页数",
          "required" : false,
          "type" : "integer",
          "default" : 0,
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "每页加载量",
          "required" : false,
          "type" : "integer",
          "default" : 10,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ "后台管理权限接口" ],
        "summary" : "创建后台管理权限",
        "description" : "创建一个后台管理权限",
        "operationId" : "createSysAuthority",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "后台管理权限",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SysAuthorityView"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "successful operation",
            "headers" : {
              "location" : {
                "type" : "string",
                "description" : "URL of new created resource"
              }
            }
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ "后台管理权限接口" ],
        "summary" : "删除后台管理权限",
        "description" : "批量删除后台管理权限",
        "operationId" : "deleteSysAuthoritys",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "后台管理权限ids",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/api/v1/sysauthorities/{id}" : {
      "get" : {
        "tags" : [ "后台管理权限接口" ],
        "summary" : "获取后台管理权限",
        "description" : "通过id获取后台管理权限",
        "operationId" : "getSysAuthority",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后台管理权限id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "put" : {
        "tags" : [ "后台管理权限接口" ],
        "summary" : "更新后台管理权限",
        "description" : "更新后台管理权限信息",
        "operationId" : "updateSysAuthority",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "权限id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "权限信息",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/SysAuthorityView"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful"
          },
          "404" : {
            "description" : "not found"
          },
          "409" : {
            "description" : "conflict"
          },
          "500" : {
            "description" : "internal Server Error"
          }
        }
      },
      "delete" : {
        "tags" : [ "后台管理权限接口" ],
        "summary" : "删除后台管理权限",
        "description" : "通过id删除后台管理权限",
        "operationId" : "deleteSysAuthority",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "后台管理权限id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SysAuthorityView" : {
      "type" : "object",
      "required" : [ "createTime", "updateTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "updateTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "enabled" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditors" : {
          "type" : "string"
        },
        "auditStatus" : {
          "type" : "integer",
          "format" : "int32"
        },
        "auditContent" : {
          "type" : "string"
        },
        "auditTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "hasAuthority" : {
          "type" : "boolean",
          "default" : false
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
...
  }
}